refactor: move passwordinput.tsx outside of ui folder
